testeado costos en tarea

+++ src/main/kotlin/ar/edu/unahur/obj2/tareas/Empleados.kt
@@ -1,5 +1,5 @@
 package ar.edu.unahur.obj2.tareas
 
-open class Empleados(val cuantoCobraPorHora :Int) {
+open class Empleados(val cuantoCobraPorHora :Double) {
     fun cuantoCobraUnEmpleadoPorHora() = cuantoCobraPorHora
 }
